% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eppo_token.R
\name{create_eppo_token}
\alias{create_eppo_token}
\title{EPPO Data Services token variable}
\usage{
create_eppo_token(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Global variable with the token to access EPPO Data Services API.
}
\description{
\code{create_eppo_token} makes global variable eppo_token. It contains token
of class string, that is needed to communicate with EPPO Data Services API.
Global variable is shared between package functions.
}
\seealso{
To obtain your free EPPO token please register
\url{https://data.eppo.int/}
}
