% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppo_citation_helpers.R
\name{eppo_citation}
\alias{eppo_citation}
\title{Citation helper tools}
\usage{
eppo_citation(cite, output = NULL)
}
\arguments{
\item{cite}{string. Possible values: 'data_services', 'global_database' or
'both'.}

\item{output}{string. Name of the file to which output will be saved. If NULL
(default) results will be printed to console.}
}
\value{
citation of EPPO Data Services, EPPO Global Database or both in
bibtex format.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#experimental'><img src='figures/lifecycle-experimental.svg' alt='Experimental lifecycle'></a>}}{\strong{Experimental}}
Citation returns citation of EPPO Data Services and/or EPPO Global Database
in bibtex format. To save the output into file use \code{output} argument
with destination file name.
}
\examples{
eppo_citation("data_services")
eppo_citation("global_database")
eppo_citation("both")
}
