% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppo_database_tools.R
\name{eppo_database}
\alias{eppo_database}
\alias{eppo_database_check}
\alias{eppo_database_download}
\alias{eppo_database_connect}
\title{EPPO Data Services Database connection tools}
\usage{
eppo_database_check(filepath = getwd(), filename = "eppocodes.sqlite")

eppo_database_download(filepath = getwd())

eppo_database_connect(filepath = getwd(), filename = "eppocodes.sqlite")
}
\arguments{
\item{filepath}{A string with path where eppocodes.sqlite file is stored. By
default it points to working directory.}

\item{filename}{A string with name of the file. By default it takes name
eppocodes.sqlite.}
}
\value{
Checks if database file exist in directory, if it is outdated, and
   establishes SQLite database connection
}
\description{
\code{eppo_database_check} checks if there is a file \emph{eppocodes.sqlite}
and informs user if it is outdated and should be downloaded.
\code{eppo_database_download} downloads database in SQLite format directly.
\code{eppo_databse_connect} allows user to conect to SQLite database
downloaded from EPPO Data Services.
}
