% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppo_full_table.R
\name{eppo_table_full}
\alias{eppo_table_full}
\title{EPPO Data Services information table}
\usage{
eppo_table_full(names_vector, sqlConnection, token)
}
\arguments{
\item{names_vector}{A vector with organism names (or part of the names) to be
checked for existence in EPPO SQLite Database.}

\item{sqlConnection}{connection to SQLite EPPO Database. By default NULL;
function will automatically connect to database with default credentials.}

\item{token}{An object containing EPPO API token created via
{\link{create_eppo_token}}.}
}
\value{
Wraps results of function from \code{eppo_tabletools}, namely:
{\link{eppo_tabletools_names}}, {\link{eppo_tabletools_hosts}},
{\link{eppo_tabletools_cat}}, {\link{eppo_tabletools_taxo}}, and
{\link{eppo_tabletools_distri}} into one compact, human readable table.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
\code{eppo_table_full} creates table in human friendly format with data on
pest preferred name, synonyms and common names, hosts, categorization,
taxonomy and its distribution.
}
