% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_eppo_token.R
\name{create_eppo_token}
\alias{create_eppo_token}
\title{EPPO Data Services token variable}
\usage{
create_eppo_token(x)
}
\arguments{
\item{x}{A string.}
}
\value{
Token to access EPPO Data Services API in format recognized by
\code{eppo_tabletools_*} functions.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
\code{create_eppo_token} makes token thats should be saved in a variable and
used argument for \code{\link{eppo_tabletools_hosts}},
\code{\link{eppo_tabletools_pests}}, \code{\link{eppo_tabletools_cat}} and
\code{\link{eppo_tabletools_taxo}} functions. It contains token of class
string, that is needed to communicate with EPPO Data Services API.
}
\examples{
\dontrun{
  create_eppo_token("12345678abcdef")
}
}
\seealso{
To obtain your free EPPO token please register
\url{https://data.eppo.int/}
}
