% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppo_tabletools.R
\name{eppo_tabletools}
\alias{eppo_tabletools}
\alias{eppo_tabletools_names}
\alias{eppo_tabletools_hosts}
\alias{eppo_tabletools_cat}
\alias{eppo_tabletools_taxo}
\alias{eppo_tabletools_distri}
\alias{eppo_tabletools_pests}
\title{EPPO table manipulation tools}
\usage{
eppo_tabletools_names(names_tables)

eppo_tabletools_hosts(
  names_tables = NULL,
  token,
  raw_eppocodes = NULL,
  use_raw_codes = FALSE
)

eppo_tabletools_cat(
  names_tables = NULL,
  token,
  raw_eppocodes = NULL,
  use_raw_codes = FALSE
)

eppo_tabletools_taxo(
  names_tables = NULL,
  token,
  raw_eppocodes = NULL,
  use_raw_codes = FALSE
)

eppo_tabletools_distri(
  names_tables = NULL,
  raw_eppocodes = NULL,
  use_raw_codes = FALSE
)

eppo_tabletools_pests(
  names_tables = NULL,
  token,
  raw_eppocodes = NULL,
  use_raw_codes = FALSE
)
}
\arguments{
\item{names_tables}{A list of tables created via {\link{eppo_names_tables}}.}

\item{token}{An object containing EPPO API token created via
{\link{create_eppo_token}}.}

\item{raw_eppocodes}{A character vector of eppocodes. Use with caution, and ONLY
when sure that provided eppocodes are correct.}

\item{use_raw_codes}{logical. Default FALSE. Set TRUE if you want to provide
eppocodes directly.}
}
\value{
List containing two data frames. First is in a long format, and each
row contains synonyms and names in other languages in respect to preferred
names. The second data frame contains coerced synonyms and other names to
single cell for each preferred name.
}
\description{
\ifelse{html}{\out{<a href='https://www.tidyverse.org/lifecycle/#maturing'><img src='figures/lifecycle-maturing.svg' alt='Maturing lifecycle'></a>}}{\strong{Maturing}}
\code{eppo_tabletools_names} creates table with names -- preferred, common
and synonyms -- of pests. It is particularly useful for creating whole table
with {\link{eppo_table_full}}, otherwise it might be more informative to use
{\link{eppo_names_tables}} which provides additional information for your
query.
\code{eppo_tabletools_hosts} creates table with hosts of pests.
\code{eppo_tabletools_cat} creates table with categorization of pests.
\code{eppo_tabletools_taxo} creates table with taxonomy of pests.
\code{eppo_tabletools_distri} creates table with distribution of pests --
continent and country level only.
\code{eppo_tabletools_pests} creates table with pests of host.
All functions return both long tables and compact, human friendly tables.
}
