% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eppo_tabletools.R
\name{eppo_tabletools}
\alias{eppo_tabletools}
\alias{eppo_tabletools_names}
\alias{eppo_tabletools_hosts}
\title{EPPO table manipulation tools}
\usage{
eppo_tabletools_names(names_tables)

eppo_tabletools_hosts(names_tables, token = eppo_token)
}
\arguments{
\item{names_tables}{A list of tables created via {\link{eppo_names_tables}}.}
}
\value{
List containing two data frames. First is in a long format, and each
  row contains synonyms and names in other languages in respect to preferred
  names. The second data frame contains coerced synonyms and other names to
  single cell for each preferred name.
}
\description{
\code{eppo_tabletools_names} creates table with names.
}
