% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pestr.R
\docType{package}
\name{pestr}
\alias{pestr}
\title{pestr: A package for retrieving data from
EPPO Data Services \url{https://data.eppo.int/}}
\description{
The pestr package provides function for downloading SQLite EPPO database,
searching for pest names in SQLite database, and retrieving all necessary
data from EPPO Data Services API for selected pest names.
}
